backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
lunarcalendar==0.0.9
gunicorn==21.2.0
psycopg2-binary==2.9.9

